requirements.txt
Django=4.1.10
djongo=1.3.6
django-cors-headers=4.2.0
django-rest-knox=4.2.0
djangorestframework=3.14.0
image=1.5.33
imageio=2.22.4
ortools=9.6.2534
pymongo=3.12.3
haversine
numpy
pandas
pickle